requirements.txt
# Flask and Web Serving
flask
gunicorn

# Google API/Auth for Drive Access
google-api-python-client
google-auth
google-auth-oauthlib  # <-- CORRECTED: This is the full, required package name

# Computer Vision and OCR
opencv-python-headless
numpy
pytesseract
Pillow